+++ src/Viewer/PickHandler.cpp
@@ -297,7 +297,7 @@ bool PickHandler::handleKeyDown( const osgGA::GUIEventAdapter& ea, GUIActionAdap
 		if ( isCtrlPressed ) {
 			layout->pause();
 			coreGraph->setNodesFreezed( true );
-			layout->getAlg()->setMaxDistance( distance * ( float )0.8 );
+			layout->getAlg()->setMaxDistance( distance * 0.8f );
 			coreGraph->scaleGraphToBase();
 			coreGraph->setNodesFreezed( false );
 			layout->play();
@@ -305,7 +305,7 @@ bool PickHandler::handleKeyDown( const osgGA::GUIEventAdapter& ea, GUIActionAdap
 		else {
 			layout->pause();
 			coreGraph->setNodesFreezed( true );
-			layout->getAlg()->setMaxDistance( distance * ( float )1.2 );
+			layout->getAlg()->setMaxDistance( distance * 1.2f );
 			coreGraph->scaleGraphToBase();
 			coreGraph->setNodesFreezed( false );
 			layout->play();
@@ -430,10 +430,10 @@ bool PickHandler::handleKeyDown( const osgGA::GUIEventAdapter& ea, GUIActionAdap
 		float distance = layout->getAlg()->getMaxDistance();
 
 		if ( isCtrlPressed ) {
-			layout->getAlg()->setMaxDistance( distance * ( float )0.8 );
+			layout->getAlg()->setMaxDistance( distance * 0.8f );
 		}
 		else {
-			layout->getAlg()->setMaxDistance( distance * ( float )1.2 );
+			layout->getAlg()->setMaxDistance( distance * 1.2f );
 		}
 	}
